+++ src/Common/src/TypeSystem/Canon/GenericParameterDesc.Canon.cs
@@ -1,20 +1,22 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
 using Debug = System.Diagnostics.Debug;
 
-namespace Internal.TypeSystem.Ecma
+namespace Internal.TypeSystem
 {
-    // Implements canonicalization of ECMA generic parameters 
-    public partial class EcmaGenericParameter
+    // Implements canonicalization of generic parameters 
+    public partial class GenericParameterDesc
     {
+        // TODO: seal the override
         public override bool IsCanonicalSubtype(CanonicalFormKind policy)
         {
             Debug.Assert(false, "IsCanonicalSubtype of an indefinite type");
             return false;
         }
 
+        // TODO: seal the override
         protected override TypeDesc ConvertToCanonFormImpl(CanonicalFormKind kind)
         {
             Debug.Assert(false, "ConvertToCanonFormImpl for an indefinite type");
